LayoutTests/platform/glib/imported/w3c/web-platform-tests/xhr/access-control-sandboxed-iframe-denied-without-wildcard-expected.txt

CONSOLE MESSAGE: Origin null is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: XMLHttpRequest cannot load http://www1.web-platform.test:8800/xhr/resources/pass.txt?pipe=header(Cache-Control,no-store)|header(Content-Type,text/plain)header(Access-Control-Allow-Credentials,true)|header(Access-Control-Allow-Origin,http://web-platform.test:8800) due to access control checks.


PASS Check that path exists and is accessible via CORS XHR request
PASS Sandboxed iframe is denied CORS access to server that allows parent origin
